bakugan se logro conectat
el back end esta conectado solo falta rellenarlo esperan do en los demas
